report_generation/requirements.txt
flask
google-cloud-storage
google-cloud-aiplatform
reportlab
python-dotenv
gunicorn